URL Shortener/requirements.txt
asgiref==3.3.1
Django==3.1.7
djangorestframework==3.12.2
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
